From Mike Morrin: eliminates the global variable for tracking which nibble is to be decoded by taking advantage of the fact that half octet IEs always occur in pairs, and thus a pair can be grouped together for decoding.
There was probably also some confusion caused by the macros UPPER_NIBBLE and
LOWER_NIBBLE because the GSM bit numbering is opposite to Wireshark internal
numbering, so I have changed these to be LEFT_NIBBLE and RIGHT_NIBBLE, which
corresponds to the display format in Wireshark.

The dissection order of half octet IEs has been adjusted where necessary to
align with the ordering shown in the GSM specifications.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6658

git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@40157 f5534014-38df-0310-8fa8-9805f1628bb7