Refer to a PID of 0x0002 as the PID for the HP Teaming heartbeat protocol.
Point to an HP document o the protocol, and discuss the two ways to hook
up the dissector.

Don't bother checking the destination address - either the OUI/PID
suffices, or the MAC address is both necessary and, presumably,
sufficient and we need to introduce a heuristic dissector table for SNAP
frames.  What's more, there's no guarantee that the destination address
is a MAC address - it might be absent, e.g. because you're capturing on
the Linux "any" device and are getting the "Linux cooked" header, with
only a destination address.

Don't put the source address into the Info column - it's already in the
source column *if* it exists (which it might not, for packets captured
from the "any" device and sent by the machine doing the capturing).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@40195 f5534014-38df-0310-8fa8-9805f1628bb7