From Gaurav Tungatkar:
support extended ICMP for multi-part messages as specified by
RFC 4884. Additionally, it adds support for the Interface Information Object
extension to ICMP as specified by RFC 5837.
1. RFC 4884 - This redefines certain ICMP messages to support multi-part
messages with extensions. It adds a length field to the ICMP header and defines
an extension structure to define extension objects that can be appended to the
selected ICMP messages. It retains backward compatibility with messages that do
not support the extensions defined by this RFC. 
2. RFC 5837 - This defines an extension object to identify the interface and
next-hop information to selected ICMP messages.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6632

git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@40298 f5534014-38df-0310-8fa8-9805f1628bb7