Second try. This time pipes and stdin are supported.
Use a global list containing all interfaces and only change
properties of the entries when changes are made in the GUI.
Do not misuse the list of interfaces specified on the command
line anymore.

This patch does not provide any new functionality, it just
provides the base for future extensions like removing
remote interface, mulitple airpcap devices and multiple
pipes.

This patch was provided by Irene Ruengeler.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@40693 f5534014-38df-0310-8fa8-9805f1628bb7