This should fix the buildbots...
The problem was that when reading a .pcap file, we don't have any IDBs.
If reqested to write out an pcapng file, we (now) build a dummy IDB which
uses the file's encapsulation as the interface encapsulation. Therefore
it can't be per=packet.
We need to fix this by using wtap_dump_open_ng()...


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@41122 f5534014-38df-0310-8fa8-9805f1628bb7