Add a presence flag field to the packet information structure filled in by Wiretap, to indicate whether certain fields in that structure actually have data in them.
Use the "time stamp present" flag to omit showing time stamp information
for packets (and "packets") that don't have time stamps; don't bother
working very hard to "fake" a time stamp for data files.

Use the "interface ID present" flag to omit the interface ID for packets
that don't have an interface ID.

We don't use the "captured length, separate from packet length, present"
flag to omit the captured length; that flag might be present but equal
to the packet length, and if you want to know if a packet was cut short
by a snapshot length, comparing the values would be the way to do that.

More work is needed to have wiretap/pcapng.c properly report the flags,
e.g. reporting no time stamp being present for a Simple Packet Block.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@41185 f5534014-38df-0310-8fa8-9805f1628bb7