Get rid of the error code argument to scan_local_interfaces(); nobody uses it once it's filled in.
From Evan Huus: in scan_local_interfaces(), pass NULL to
capture_interface_list(), as we don't use the error string (and don't
free it, either).

In fill_capture_box(), for CANT_GET_INTERFACE_LIST, include the error
string in the report, and free it, in all cases, when we're done with
it.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@42089 f5534014-38df-0310-8fa8-9805f1628bb7