Partial fix for https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7221 (emem alignment problems on SPARC) :
Add the room for the pointer to the next (from r31577) *before* calculating
the canary+pad: that way the complete allocation
(allocation+canary_ptr+canary+pad) will end on an 8-byte boundary (as was the
case before r31577).

This only solves the alignment problem when using canaries (i.e., not, by
default, se_ allocations).

(And, yes, this is ignoring the 16-byte alignment requirements of long
doubles.)


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@42407 f5534014-38df-0310-8fa8-9805f1628bb7