The rest of the fix for https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7221 (emem alignment problems on SPARC) :
Have emem use 8-byte alignment when we need it.

Since I can't seem to write code that which reliably (across GCC versions and
optimization levels) determines if 8-byte alignment is needed for doubles,
"when" is defined as "if we're compiling for a CPU other than i386."

Windows doesn't need a check because it's either i386 or 64-bit (x86_64 or
maybe ia64--both of which get 8-byte alignment from G_MEM_ALIGN).

(And, yes, all of this is ignoring the 16-byte alignment requirements of long
doubles.)


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@42431 f5534014-38df-0310-8fa8-9805f1628bb7